finish the algorithm of fall down

+++ android/app/src/main/java/com/example/jakeoneim/fitec/HeartRateCheck.java
@@ -16,7 +16,8 @@ public class HeartRateCheck extends Thread{
         this.em = new Emergency();
 
     }
-    public boolean isProblem(int currentHertRate){ // if there is problem true
+
+    public boolean isProblem(double currentHertRate){ // if there is problem true
         if(isProbremOccured) return false;
         if(currentHertRate < standardHeartRateOfUser/2 || currentHertRate > standardHeartRateOfUser +40){
             this.numberOfProblemChecked++;
